+++ Mimus/Source/Matchers/InstanceOfMatcher.swift
@@ -5,9 +5,9 @@
 import Foundation
 
 public final class InstanceOf<T>: MockEquatable {
-    
+
     public init() { }
-    
+
     public func equalTo(other: Any?) -> Bool {
         return other is T
     }